// src/modules/psychologist/enums/specialities.enum.ts
export enum EPsychologistSpecialty {
  ANXIETY_DISORDER = 'anxiety_disorder',
  COUPLES_THERAPY = 'couples_therapy',
  EATING_DISORDER = 'eating_disorder',
  BIPOLAR_DISORDER = 'bipolar_disorder',
  LIFE_TRANSITIONS = 'life_transitions',
  CHILD_ADOLESCENT_THERAPY = 'child_adolescent_therapy',
  SLEEP_DISORDERS = 'sleep_disorders',
  DEPRESSION = 'depression',
  FAMILY_THERAPY = 'family_therapy',
  ADHD = 'adhd',
  OCD = 'ocd',
  CAREER_COUNSELING = 'career_counseling',
  GERIATRIC_PS = 'geriatric_psychology',
  ANGER_MANAGEMENT = 'anger_management',
  TRAUMA_PTSD = 'trauma_ptsd',
  ADDICTION_SUBSTANCE_ABUSE = 'addiction_substance_abuse',
  AUTISM_SPECTRUM_DISORDER = 'autism_spectrum_disorder',
  GRIEF_LOSS = 'grief_loss',
  LGBTQIA = 'lgbtqia',
  CHRONIC_PAIN_MANAGEMENT = 'chronic_pain_management',
}