// src/modules/users/enums/insurance_accepted .enum.ts
export enum EInsurance {
  OSDE = 'osde',
  SWISSMEDICAL = 'swiss-medical',
  IOMA = 'ioma',
  PAMI = 'pami',
  UNIONPERSONAL = 'unión-personal',
  OSDEPYM = 'osdepy',
  LUISPASTEUR = 'luis-pasteur',
  JERARQUICOSSALUD = 'jerarquicos-salud',
  SANCORSALUD = 'sancor-salud',
  OSECAC = 'osecac',
  OSMECONSALUD = 'osmecón-salud',
  APROSS = 'apross',
  OSPRERA = 'osprera',
  OSPAT = 'ospat',
  SENACIONAL = 'ase-nacional',
  OSPSIP = 'ospsip',
}